requirements.txt
Django==5.0.0
django-mongodb-backend==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
